Minor changes in configuration, integration and repository layers

+++ src/main/java/com/levita/levita_monitoring/repository/UserKpiRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserKpiRepository extends JpaRepository<UserKpi, Integer> {
+public interface UserKpiRepository extends JpaRepository<UserKpi, Long> {
 
     Optional<UserKpi> findByUser(User user);
